src/Fileloader/requirements.txt
numpy==1.26.1
pdf2image==1.16.3
Pillow==10.1.0
watchdog==3.0.0
